#1176 move path only used in ch24 to ch24
