+++ src/ch24_belief_dst/vow_db2df.py
@@ -14,15 +14,12 @@ from ch17_idea.idea_belief_csv import (
     create_init_belief_idea_csv_strs,
 )
 from ch17_idea.idea_db_tool import csv_dict_to_excel, prettify_excel_file
-from ch18_etl_config._ref.ch18_path import (
-    create_belief0001_path,
-    create_moment_mstr_path,
-    create_world_db_path,
-)
+from ch18_etl_config._ref.ch18_path import create_moment_mstr_path, create_world_db_path
 from ch18_etl_config.etl_sqlstr import (
     create_prime_tablename as prime_tbl,
     create_sound_and_heard_tables,
 )
+from ch24_belief_dst._ref.ch24_path import create_belief0001_path
 from os.path import exists as os_path_exists
 from sqlite3 import Cursor as sqlite3_Cursor, connect as sqlite3_connect
