+++ src/ref/keg_definitions.json
@@ -500,7 +500,7 @@
   "Tuesday": "Example key with value 'Tuesday' used for tests throughout codebase.", 
   "unknown_str": ", TRANSLATE_LABEL arg, TRANSLATE_NAME arg, TRANSLATE_ROPE arg, TRANSLATE_TITLE arg", 
   "UPDATE": "CRUD_COMMAND to update records for a table. 'Associated dimens: personunit' in 'CRUD_COMMAND to update records for a table'", 
-  "vow": "short code for audience created beliefs", 
+  "vow": "beliefs created for audience", 
   "wed": "Example key with value 'Wed' used for tests throughout codebase.", 
   "Wednesday": "Example key with value 'Wednesday' used for tests throughout codebase.", 
   "week": "A fundamental structure of time for Keg2. Always assumed to be a days long. Never partial days.",
